+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## latest
 
+* 2020-11-08, Use ListView for flex items.
 * 2020-11-08, Refine ListTile ui, no padding.
 * 2020-11-08, Disable play button, if url is invalid.
 * 2020-11-08, Refine button, check url before play.